+++ src/ts/components/core/slider/Slider.tsx
@@ -65,6 +65,8 @@ interface Props
     restrictToMarks?: boolean;
     /** Function to generate scale (See https://www.dash-mantine-components.com/functions-as-props) A transformation function to change the scale of the slider */
     scale?: any;
+    /** Domain of the slider, defines the full range of possible values, `[min, max]` by default */
+    domain?: [number, number];
 }
 
 /** Slider */