+++ src/ts/props/dates.ts
@@ -33,6 +33,8 @@ interface CalendarHeaderSettings {
     withPrevious?: boolean;
     /** Component size */
     size?: MantineSize;
+    /** Controls order, `['previous', 'level', 'next']`` by default */
+    headerControlsOrder?: ('next' | 'previous' | 'level')[];
 }
 
 interface ControlsGroupSettings {
@@ -58,6 +60,8 @@ interface YearsListSettings extends ControlsGroupSettings {
     size?: MantineSize;
     /** Determines whether controls should be separated by spacing, true by default */
     withCellSpacing?: boolean;
+    /** A function that passes props down to year picker control based on date. (See https://www.dash-mantine-components.com/functions-as-props) */
+    getYearControlProps?: any;
 }
 
 interface DecadeLevelBaseSettings extends YearsListSettings {
@@ -74,6 +78,8 @@ interface MonthsListSettings extends ControlsGroupSettings {
     monthsListFormat?: string;
     /** Determines whether controls should be separated by spacing, true by default */
     withCellSpacing?: boolean;
+    /** A function that passes props down month picker control based on date. (See https://www.dash-mantine-components.com/functions-as-props) */
+    getMonthControlProps?: any;
 }
 
 interface YearLevelBaseSettings extends MonthsListSettings {
@@ -106,6 +112,10 @@ interface MonthSettings {
     withCellSpacing?: boolean;
     /**Determines whether week numbers should be displayed, false by default */
     withWeekNumbers?: boolean;
+    /** A function that passes props down Day component  based on date. (See https://www.dash-mantine-components.com/functions-as-props) */
+    getDayProps?: any;
+    /** A function that controls day value rendering. (See https://www.dash-mantine-components.com/functions-as-props) */
+    renderDay?: any;
 }
 
 interface MonthLevelBaseSettings extends MonthSettings {
@@ -279,6 +289,11 @@ export interface TimePickerProps
     timeRangePresets?: GetTimeRange;
 }
 
+type DatePickerPreset = {
+    value: string | [string, string];
+    label: string;
+};
+
 interface PickerBaseProps {
     /** Picker type: range, multiple or default */
     type?: DatePickerType;
@@ -288,6 +303,8 @@ interface PickerBaseProps {
     allowDeselect?: boolean;
     /** Determines whether single year can be selected as range, applicable only when type="range" */
     allowSingleDateInRange?: boolean;
+    /** Initial displayed date */
+    defaultDate?: string;
 }
 
 export interface DatePickerBaseProps
@@ -301,6 +318,8 @@ export interface DatePickerBaseProps
     maxLevel?: CalendarLevel;
     /** Current level displayed to the user (decade, year, month), used for controlled component */
     level?: CalendarLevel;
+    /** Predefined values to pick from */
+    presets?: DatePickerPreset[];
 }
 
 export interface MonthPickerBaseProps
@@ -318,8 +337,7 @@ export interface MonthPickerBaseProps
 export interface YearPickerBaseProps
     extends PickerBaseProps,
         DecadeLevelBaseSettings,
-        CalendarBaseProps,
-        CalendarSettings {}
+        CalendarBaseProps {}
 
 export interface TimeGridProps extends BoxProps, StylesApiProps {
     /** Time data in 24h format to be displayed in the grid, for example `['10:00', '18:30', '22:00']`. Time values must be unique. */